(#136) Move FireWhen to a module
* added helper folder for FireWhen.js
* Replaced instances of window.FireWhen to instead call new FireWhen module

Closes #136
